# set default value for wal_bytes_per_sync and bytes_per_sync