After i have test, when sync_log = false, 512KB is a appropriate value for wal_bytes_per_sync. It doesn't require as high IOPS as 256KB and 128KB.
RocksDB will call fdatasync before sst file closed. We set bytes_per_sync = 1MB to make fsync more smoothly.