@MyonKeminta 2018-03-16
## Command line:

```bash
# master:  bytes_per_sync=0,       wal_bytes_per_sync=0
# this pr: bytes_per_sync=1048576, wal_bytes_per_sync=524288

./db_bench \
  --db=./test_db \
  --use_existing_db=false \
  --key_size=30 \
  --value_size=256 \
  --bloom_bits=10 \
  --benchmarks=fillrandom \
  --memtablerep=skip_list \
  --compression_type=Snappy \
  --compression_ratio=0.5 \
  --num=10000000 \
  --bytes_per_sync=$bytes_per_sync \
  --wal_bytes_per_sync=$wal_bytes_per_sync
```

## Result of db_bench: 

(test two configurations in turn)

```
master:   4.014 micros/op 249114 ops/sec;   67.9 MB/s
this pr:  3.937 micros/op 253987 ops/sec;   69.3 MB/s
master:   3.873 micros/op 258230 ops/sec;   70.4 MB/s
this pr:  3.998 micros/op 250131 ops/sec;   68.2 MB/s
master:   3.948 micros/op 253264 ops/sec;   69.1 MB/s
this pr:  4.024 micros/op 248502 ops/sec;   67.8 MB/s
master:   3.840 micros/op 260421 ops/sec;   71.0 MB/s
this pr:  3.920 micros/op 255122 ops/sec;   69.6 MB/s
master:   3.901 micros/op 256314 ops/sec;   69.9 MB/s
this pr:  4.054 micros/op 246695 ops/sec;   67.3 MB/s
```

## Result of iostat:

### Command Line:

```bash
iostat nvme1n1 -m -x 1
```

### Partial records of master

```
avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           3.10    0.00    1.80    0.05    0.00   95.05

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 22839.00    0.00 1254.00     0.00   151.75   247.84     1.42    1.13    0.00    1.13   0.06   8.00

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           3.33    0.00    1.73    0.05    0.00   94.90

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 21897.00    0.00 1547.00     0.00   190.06   251.61     3.70    2.39    0.00    2.39   0.07  10.10

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.58    0.00    2.05    0.08    0.00   93.30

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 25550.00    0.00 1482.00     0.00   180.75   249.79     2.80    1.89    0.00    1.89   0.06   9.40

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.20    0.00    2.25    0.08    0.00   93.47

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 51360.00    0.00 2578.00     0.00   316.05   251.08     4.32    1.67    0.00    1.67   0.06  16.40

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           3.20    0.00    1.70    0.05    0.00   95.05

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 12923.00    1.00  992.00     0.00   119.45   246.36     3.07    3.09    0.00    3.09   0.07   6.90

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.18    0.00    2.13    0.05    0.00   93.65

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 39605.00    0.00 2141.00     0.00   262.21   250.82     2.27    1.06    0.00    1.06   0.06  13.30

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.60    0.00    2.33    0.05    0.00   93.03

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 33013.00    0.00 1555.00     0.00   190.86   251.37     2.93    1.88    0.00    1.88   0.06  10.00

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.23    0.00    2.00    0.13    0.00   93.65

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 41028.00    0.00 2857.00     0.00   303.95   217.88     9.17    3.21    0.00    3.21   0.05  15.70

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           3.93    0.00    1.98    0.05    0.00   94.05

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 24062.00    0.00 1372.00     0.00   166.45   248.47     2.08    1.51    0.00    1.51   0.06   8.70

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.15    0.00    2.00    0.08    0.00   93.77

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00 41923.00    0.00 2219.00     0.00   272.07   251.10     2.96    1.33    0.00    1.33   0.07  14.50

```

### Partial records of this pr

```
avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           2.85    0.00    1.60    0.00    0.00   95.55

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  1486.00    0.00 1373.00     0.00   159.96   238.61     0.34    0.25    0.00    0.25   0.05   6.40

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           3.55    0.00    1.58    0.00    0.00   94.87

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00   685.00    0.00 1500.00     0.00   175.35   239.41     0.32    0.21    0.00    0.21   0.04   6.40

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.30    0.00    1.80    0.00    0.00   93.90

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  2241.00    1.00 1905.00     0.00   220.82   237.28     0.45    0.24    0.00    0.24   0.05   9.10

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.30    0.00    1.78    0.03    0.00   93.90

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  2479.00    0.00 2021.00     0.00   234.32   237.45     0.40    0.20    0.00    0.20   0.04   8.20

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           2.98    0.08    1.63    0.00    0.00   95.32

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  1923.00    0.00 1369.00     0.00   158.89   237.70     0.28    0.20    0.00    0.20   0.05   6.70

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.20    0.00    1.85    0.00    0.00   93.95

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  1940.00    0.00 1955.00     0.00   227.81   238.65     0.37    0.19    0.00    0.19   0.04   7.80

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.05    0.00    1.85    0.00    0.00   94.10

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  1990.00    0.00 2009.00     0.00   232.80   237.32     0.45    0.22    0.00    0.22   0.04   8.90

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.33    0.00    2.00    0.00    0.00   93.67

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  2253.00    1.00 2148.00     0.00   248.26   236.59     0.52    0.24    0.00    0.24   0.05  10.20

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.30    0.00    2.03    0.03    0.00   93.65

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  2448.00    0.00 2143.00     0.00   246.53   235.60     0.46    0.21    0.00    0.21   0.04   9.50

avg-cpu:  %user   %nice %system %iowait  %steal   %idle
           4.25    0.00    1.93    0.00    0.00   93.82

Device:         rrqm/s   wrqm/s     r/s     w/s    rMB/s    wMB/s avgrq-sz avgqu-sz   await r_await w_await  svctm  %util
nvme1n1           0.00  1686.00    0.00 1917.00     0.00   222.81   238.04     0.37    0.19    0.00    0.19   0.04   8.30

```

## Conclusions

* No significant effect on OPS are shown
* wrqm/s [Write requests (before merge) queued to device per second] reduced by about 90%
* Variance of w/s [Write requests (after merge) completed per second] reduced
* iostat recorded a tiny value of iowait of CPU when testing configuration of master, but this tiny value almost disappeared in this pr's configuration